fix(cantidad): store as int not float, show as Nx format
- Change float(qty_dec) to int(qty_dec) in views.py (source of '1,0' bug)
- Display {{ item.cantidad|floatformat:0 }}x in imprimir.html for both
  llevar and mesa items

+++ facturacion/views.py
@@ -237,7 +237,7 @@ class FacturaLlevarCreateView(ClienteScopeMixin, PermissionRequiredMixin, LoginR
                 items.append({
                     "producto_id": producto.pk,
                     "producto_nombre": producto.nombre,
-                    "cantidad": float(qty_dec),
+                    "cantidad": int(qty_dec),
                     "precio_unitario": float(price_dec),
                     "subtotal": float(item_subtotal),
                 })
